requirements.txt
aniso8601==8.1.1; python_version >= "3.7" and python_version < "4.0"
anyio==3.6.1; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.2"
bcrypt==4.0.0; python_version >= "3.6"
certifi==2022.9.14; python_version >= "3.7" and python_version < "4" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0")
charset-normalizer==2.1.1; python_version >= "3.7" and python_version < "4" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0") and python_full_version >= "3.6.0"
click==8.1.3; python_version >= "3.7"
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0" and platform_system == "Windows"
evernote3 @ git+https://github.com/evernote/evernote-sdk-python3.git@master
fastapi==0.85.0; python_version >= "3.7"
graphene-mongo==0.2.14; python_version >= "3.6"
graphene==3.0b7; python_version >= "3.7" and python_version < "4.0"
graphql-core==3.2.1; python_version >= "3.7" and python_version < "4"
graphql-relay==3.2.0; python_version >= "3.7" and python_version < "4"
h11==0.13.0; python_version >= "3.7"
httplib2==0.20.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
httptools==0.5.0; python_version >= "3.7" and python_full_version >= "3.5.0"
idna==3.4; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.2" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0")
iso8601==1.0.2; python_full_version >= "3.6.2" and python_version < "4.0" and python_version >= "3.6"
loguru==0.6.0; python_version >= "3.5"
mongoengine==0.24.2; python_version >= "3.6"
oauth2==1.9.0.post1
oauthlib==3.2.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
promise==2.3; python_version >= "3.6"
pydantic==1.10.2; python_version >= "3.7"
pyjwt==2.4.0; python_version >= "3.6"
pymongo==4.2.0; python_version >= "3.7"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version > "3.0"
python-dotenv==0.17.1
pyyaml==6.0; python_version >= "3.7"
requests-oauthlib==1.3.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
requests==2.28.1; python_version >= "3.7" and python_version < "4" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0")
singledispatch==3.7.0; python_version >= "3.6"
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
sniffio==1.3.0; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.2"
starlette-graphene3==0.6.0; python_version >= "3.7" and python_version < "4.0"
starlette==0.20.4; python_version >= "3.7" and python_version < "4.0"
typing-extensions==4.3.0; python_version < "3.10" and python_version >= "3.7"
urllib3==1.26.12; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0") or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0")
uvicorn==0.18.3; python_version >= "3.7"
uvloop==0.17.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.7"
watchfiles==0.17.0; python_version >= "3.7"
websockets==10.3; python_version >= "3.7"
win32-setctime==1.1.0; sys_platform == "win32" and python_version >= "3.5"
